Decompose config file
