+++ scripts/file-length-allowlist.txt
@@ -17,5 +17,4 @@
 # be split, not grandfathered.
 
 # --- production code (limit 1000) ---
-config/config.go 1128
 session/backend_hook.go 1026
